requirements.txt
pymupdf 
requests 
google-generativeai 
PyPDF2 
pdfplumber
google-cloud-aiplatform
streamlit